MDEV-11675. Embedded test failure, assert opened up and satisfied, cleanup
The test is guarded to not run on embedded.
In a case of STOP SLAVE that separates SA (has been done) and CA
(will not 'cos of the stop), see rpl_start_alter_restart_slave.test,
an assert got fired when a "inadvertently" commented-out statement has
been finally identified and restored for action.
It turns out the split alter executing parallel workers did not fully comply
with the stop slave's worker stop protocol and could be on loose
at time the SQL thread thought their have been docked.
It's fixed with code relocation and *removal* an unnecessary
complicating if.