MDEV-14425 Improve the redo log for concurrency
The InnoDB redo log used to be formatted in blocks of 512 bytes.
The log blocks were encrypted and the checksum was calculated while
holding log_sys.mutex, creating a serious scalability bottleneck.

We remove the fixed-size redo log block structure altogether and
essentially turns every mini-transaction into a log block of its own.
This allows encryption and checksum calculations to be performed
in mtr_t::m_log, before acquiring log_sys.mutex. The mutex only
protects a memcpy() of the data to the shared log_sys.buf, as well
as the padding of the log, in case the to-be-written part of the log
would not end in a block boundary of the underlying storage.

We will require ib_logfile0 to exist; the redo log will not be
automatically created if it is missing. Only with with
innodb_force_recovery=6, it is possible to start InnoDB in read-only
mode even if the log file does not exist. This allows the contents
of a possibly corrupted database to be dumped.

We extend the MDEV-12353 redo log record format as follows.

(1) Empty mini-transactions or extra NUL bytes will not be allowed.
(2) The end-of-minitransaction marker (a NUL byte) will be replaced
with a 1-bit sequence number, which will be toggled each time when the
circular log file wraps back to the beginning.
(3) After the sequence bit, a CRC-32C checksum of all data
(excluding the sequence bit) will written.
(4) If the log is encrypted, 8 bytes will be written after
the checksum. This is part of the initialization vector (IV)
of encrypted log data.
(5) File names, page numbers, and checkpoint information will not be
encrypted. Only the payload bytes of page-level log will be encrypted.
The tablespace ID and page number will form part of the IV.
(6) For padding, arbitrary-length FILE_CHECKPOINT records may be written,
with all-zero payload, and with the normal end marker and checksum.
The log_pad() function has not been enabled yet.

The 512-byte log checkpoint blocks at 0x200 and 0x600 will be replaced
with 64-byte log checkpoint blocks at 0x1000 and 0x2000.

The start of log records will move from 0x800 to 0x3000. This allows us
to use 4096-byte aligned blocks for all I/O in a future revision.

In mariadb-backup, we will no longer remove ib_logfile0 or create an
empty ib_logfile0. Server startup will require a valid log file.
When the log is resized, we will create a logically empty ib_logfile101
and use an atomic rename to replace ib_logfile0 with it.

Because there is no mandatory padding in the log file, we are able
to create a dummy log file as of an arbitrary log sequence number.
See the test mariabackup.huge_lsn.

An attempt to start up the server without a valid ib_logfile0 will be
refused. See the test innodb.log_file_size.  When resizing the log,
we will first initialize ib_logfile101 and then atomically replace the
old ib_logfile0 with it.

The parameter innodb_log_write_ahead_size and the
INFORMATION_SCHEMA.INNODB_METRICS counter log_padded will be removed.

Thanks to Matthias Leich for providing https://rr-project.org traces
for various failures during the development, and to
Thirunarayanan Balathandayuthapani for his help in debugging
some of the recovery code. And thanks to the developers of the
rr debugger for a tool without which extensive changes to InnoDB
would be very challenging to get right.

TODO: Remove recv_sys_t::buf, and use log_sys.buf for parsing,
and test and improve the performance, especially on PMEM, where
we can directly write to log_sys.buf=mmap(MAP_SYNC).

TODO: When writing log blocks, use the file system
block size instead of using arbitrary byte offsets.
For example, write only up to the last completely filled block
(as determined by the file system and the underlying storage),
or pad the last block to full size, or just write garbage after
the current record, and then overwrite that block later.

recv_sys_t::MTR_SIZE_MAX: Replaces RECV_SCAN_SIZE.

recv_sys_t::file_checkpoint: Renamed from mlog_checkpoint_lsn.

recv_sys_t, log_sys_t: Removed many data members.

os_file_create_func() [WIN32]: Clean up the code.
No longer set FILE_FLAG_NO_BUFFERING on the log file
because for now, log file writes will not be aligned.

mariadb-backup: Eliminated several variables. We will refer
directly to recv_sys and log_sys.

backup_wait_for_lsn(): Detect non-progress of
xtrabackup_copy_logfile(). In this new log format with
arbitrary-sized blocks, we can only detect log file overrun
indirectly, by observing that the scanned log sequence number
is not advancing.